+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 1.2.0 - TBA
 
+- Fix #8 Automatically adjust datepicker position
+- Merge #45 Adjusting position
+- Merge #11 Allow using a different element to open and bind the datepicker to
+- Use new Bower package filename: bower.json
+- Merge #60 Add position alignement option (top|bottom && left|right)
+
 ## 1.1.0 - 2013-03-29
 
 - Add Ruby on Rails [gem](https://rubygems.org/gems/pikaday-gem)